更新

+++ tts.cpp/ggml-patches/ggml-iterator.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <utility>
 
 #include "ggml.h"
@@ -7,13 +8,13 @@
 
 class gguf_key_iterator {
     const gguf_context * const ctx;
-    const int                  n_kv;
-    int                        i{};
+    const int64_t              n_kv;
+    int64_t                    i{};
 
   public:
     explicit gguf_key_iterator(const gguf_context & ctx) : ctx{ &ctx }, n_kv{ gguf_get_n_kv(&ctx) } {}
 
-    std::pair<int, const char *> operator*() const { return { i, gguf_get_key(ctx, i) }; }
+    std::pair<int64_t, const char *> operator*() const { return { i, gguf_get_key(ctx, static_cast<int>(i)) }; }
 
     gguf_key_iterator & operator++() {
         ++i;
