+++ tts.cpp/src/models/loaders.cpp
@@ -1,6 +1,7 @@
 #include "loaders.h"
 
 #include <cstring>
+#include <cstdint>
 #include <unordered_map>
 
 #include "../util.h"
@@ -74,13 +75,13 @@ static unique_ptr<tts_generation_runner> runner_from_file_impl(const char * fnam
         GGML_ABORT("gguf_init_from_file failed for file %s\n", fname);
     }
     if (use_mmap) {
-        const int n{ gguf_get_n_tensors(&*meta_ctx) };
-        int       i{};
+        const int64_t n{ gguf_get_n_tensors(&*meta_ctx) };
+        int64_t       i{};
         void *    in_buffer{ static_cast<char *>(in_mmap->addr()) + gguf_get_data_offset(meta_ctx) };
         for (ggml_tensor & cur : ggml_tensor_iterator{ *weight_ctx }) {
             GGML_ASSERT(i < n);
-            GGML_ASSERT(!strcmp(cur.name, gguf_get_tensor_name(&*meta_ctx, i)));
-            cur.data = static_cast</*const*/ char *>(in_buffer) + gguf_get_tensor_offset(&*meta_ctx, i);
+            GGML_ASSERT(!strcmp(cur.name, gguf_get_tensor_name(&*meta_ctx, static_cast<int>(i))));
+            cur.data = static_cast</*const*/ char *>(in_buffer) + gguf_get_tensor_offset(&*meta_ctx, static_cast<int>(i));
             ++i;
         }
     }
